Refactor: extract base class from GoalError
