+++ base/src/main/java/org/arend/typechecking/visitor/DumbTypechecker.java
@@ -88,7 +88,7 @@ public class DumbTypechecker extends VoidConcreteVisitor<Void, Void> {
 
   @Override
   public Void visitGoal(Concrete.GoalExpression expr, Void params) {
-    myTypechecker.errorReporter.report(new GoalError(null, null, null, null, Collections.emptyList(), null, expr));
+    myTypechecker.errorReporter.report(new GoalError(null, Collections.emptyMap(), null, null, Collections.emptyList(), null, expr));
     return null;
   }
